packages/ai-engine/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data & ML
pydantic==2.5.3
numpy==1.26.3
pandas==2.1.4
scikit-learn==1.4.0
joblib==1.3.2

# Database
supabase==2.3.4
python-dotenv==1.0.0

# CORS & Security
python-jose[cryptography]==3.3.0

# Optional: Deep Learning (uncomment if needed)
# tensorflow==2.15.0
# torch==2.1.2
